仮想曲線ライントレースの実装(CurveVirtual)
updateで値更新とresetの中でx,yの値を受け取るとこまで

+++ device/Brightness.cpp
@@ -6,7 +6,7 @@ Brightness::Brightness()
 
 }
 
-float Brightness::getValue()
+double Brightness::getValue()
 {
     return mLpf->getFillteredValue();
 }
