+++ device/Length.h
@@ -1,12 +1,13 @@
 #ifndef __LENGTH_H__
 #define __LENGTH_H__
+
 #include "Measure.h"
 
 class Length : public Measure
 {
     public:
         Length();
-        float getValue();
+        double getValue();
         void update(float len);
 
     private:
