+++ device/Measure.h
@@ -7,14 +7,13 @@ class Measure
 {
     public:
         Measure();
-        virtual float getValue();
+        virtual double getValue();
 
     protected:
-            LowPassFilter *mLpf;
-
+        LowPassFilter *mLpf;
+        
     private:
 
 };
-
 #endif
